Details final details

+++ rosmery-salazar-react-test/src/utilities/Columns.jsx
@@ -4,16 +4,33 @@ import DeleteIcon from '@mui/icons-material/Delete';
 import InfoIcon from '@mui/icons-material/Info';
 import { deleteData, readData } from './providers';
 import { useNavigate } from 'react-router-dom';
+import { modalCheck, modalError, modalInfo, modalSuccess } from './modals';
 
 
 export const Columns = () =>{ 
 
   const navigate = useNavigate()
 
-  const handleDelete = async (id) => {    
-    await deleteData(id)
-    readData()
+  const handleDelete = async (id) => { 
+
+    const result = await modalCheck();
+    
+    if (result.isConfirmed) {
+      try {
+        await deleteData(id);  
+        readData();     
+        modalSuccess('Deleted!', 'The product has been deleted', 'success');
+      } catch (error) {
+        
+        modalError('Error', 'There was a problema trying to delete the product.', 'error');
+      }
+    } else {
+     
+      modalInfo('Canceled', 'The action was canceled.', 'info');
+    }
+    
   };
+
   const handleEdit = async(row) => { 
     localStorage.setItem('productEdit', JSON.stringify(row));   
     navigate(`/newproduct/${row.id}`);
